+++ cas-20200630/alibabacloud_cas20200630/models/_describe_pca_and_external_cacertificate_list_request.py
@@ -8,9 +8,13 @@ class DescribePcaAndExternalCACertificateListRequest(DaraModel):
     def __init__(
         self,
         current_page: int = None,
+        identifiers: str = None,
+        key_word: str = None,
         show_size: int = None,
     ):
         self.current_page = current_page
+        self.identifiers = identifiers
+        self.key_word = key_word
         self.show_size = show_size
 
     def validate(self):
@@ -24,6 +28,12 @@ class DescribePcaAndExternalCACertificateListRequest(DaraModel):
         if self.current_page is not None:
             result['CurrentPage'] = self.current_page
 
+        if self.identifiers is not None:
+            result['Identifiers'] = self.identifiers
+
+        if self.key_word is not None:
+            result['KeyWord'] = self.key_word
+
         if self.show_size is not None:
             result['ShowSize'] = self.show_size
 
@@ -33,6 +43,12 @@ class DescribePcaAndExternalCACertificateListRequest(DaraModel):
         m = m or dict()
         if m.get('CurrentPage') is not None:
             self.current_page = m.get('CurrentPage')
+
+        if m.get('Identifiers') is not None:
+            self.identifiers = m.get('Identifiers')
+
+        if m.get('KeyWord') is not None:
+            self.key_word = m.get('KeyWord')
 
         if m.get('ShowSize') is not None:
             self.show_size = m.get('ShowSize')